LayoutTests/platform/chromium-android/fast/forms/time/time-stepup-stepdown-from-renderer-expected.txt

Check stepping-up and -down for time input fields from renderer. No cases of empty initial values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Function arguments are (value, step, {min or max}, [stepCount]).
Normal cases
PASS stepUp("20:13", null, null) is "20:14"
PASS stepDown("20:13", null, null) is "20:12"
PASS stepUp("20:13", null, null, 10) is "20:23"
PASS stepDown("20:13", null, null, 11) is "20:02"
PASS stepUp("20:13", "4", null, 2) is "20:13:08"
PASS stepDown("20:13", "4", null, 3) is "20:12:48"
Step=any
PASS stepUp("20:13", "any", null) is "20:14"
PASS stepDown("20:13", "any", null) is "20:12"
Overflow/underflow
PASS stepUp("20:13", "3.40282346e+38", null) is "23:59:59.999"
PASS stepDown("20:13", "3.40282346e+38", null) is "00:00:00"
PASS stepUp("20:13", "1", "20:13") is "20:13"
PASS stepDown("20:13", "1", "20:13") is "20:13"
PASS stepUp("23:59", null, null) is "23:59"
PASS stepDown("00:00", null, null) is "00:00"
stepDown()/stepUp() for stepMismatch values
PASS stepDown("20:13", "3", "20:12:56") is "20:12:59"
PASS stepUp("00:13", "7", "") is "00:13:04"

PASS successfullyParsed is true

TEST COMPLETE
